Split key_date class into separate key and date classes
