+++ lib/key.rb
@@ -1,8 +1,7 @@
-class KeyDate
+class Key
 
   def initialize
     @number = random_number
-    @date = date
   end
 
   def random_number
@@ -12,10 +11,4 @@ class KeyDate
     end
     num_array.join
   end
-
-  def date
-    time = Time.new
-    time.strftime("%m%d%y")
-  end
-
 end
